// src/models/IoK8sApiNetworkingV1alpha1ServiceCIDRList/types.ts
import { IoK8sApiNetworkingV1alpha1ServiceCIDRModelOptions as IoK8sApiNetworkingV1alpha1ServiceCIDR } from '../IoK8sApiNetworkingV1alpha1ServiceCIDR';
import { IoK8sApimachineryPkgApisMetaV1ListMetaModelOptions as IoK8sApimachineryPkgApisMetaV1ListMeta } from '../IoK8sApimachineryPkgApisMetaV1ListMeta';

export interface IoK8sApiNetworkingV1alpha1ServiceCIDRListModelOptions {
  apiVersion?: string;
  items: Array<IoK8sApiNetworkingV1alpha1ServiceCIDR>;
  kind?: string;
  metadata?: IoK8sApimachineryPkgApisMetaV1ListMeta;
}